# services/core/shared/generated/api/spec/openapi.json
{"openapi":"3.0.0","paths":{"/accounts/telegram":{"post":{"operationId":"createTelegramAccount","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTelegramAccountDto"}}}},"responses":{"201":{"description":"Account data","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountWithUserDto"}}}}},"summary":"Create a new telegram account","tags":["accounts"]}},"/accounts":{"get":{"operationId":"getAllAccounts","parameters":[],"responses":{"200":{"description":"List of accounts","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AccountWithUserDto"}}}}}},"summary":"Get all accounts","tags":["accounts"]}},"/accounts/{type}/{channelId}":{"get":{"operationId":"getAccount","parameters":[{"name":"type","required":true,"in":"path","schema":{"enum":["TELEGRAM","VK"],"type":"string"}},{"name":"channelId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"Account data","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountWithUserDto"}}}}},"summary":"Get account","tags":["accounts"]}},"/users":{"get":{"operationId":"getAllUsers","parameters":[],"responses":{"200":{"description":"List of users","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserDto"}}}}}},"summary":"Get all users","tags":["users"]}},"/users/{id}":{"get":{"operationId":"getUser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"Single user","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}},"summary":"Get a user","tags":["users"]},"patch":{"operationId":"updateUser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":"Updated user","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}},"summary":"Update a user","tags":["users"]},"delete":{"operationId":"removeUser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"Deleted user","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}},"summary":"Delete a user","tags":["users"]}},"/events":{"post":{"operationId":"createEvent","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateEventDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EventDto"}}}}},"summary":"Create a new event","tags":["events"]},"get":{"operationId":"getEvents","parameters":[{"name":"type","required":false,"in":"query","schema":{"enum":["COMMON","TRIP"],"type":"string"}},{"name":"upcoming","required":false,"in":"query","schema":{"type":"boolean"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/EventDto"}}}}}},"summary":"Get events","tags":["events"]}},"/events/members":{"get":{"operationId":"getEventsWithMembers","parameters":[{"name":"type","required":false,"in":"query","schema":{"enum":["COMMON","TRIP"],"type":"string"}},{"name":"upcoming","required":false,"in":"query","schema":{"type":"boolean"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/EventWithMembers"}}}}}},"summary":"Get events with members","tags":["events"]}},"/events/{id}":{"get":{"operationId":"getEvent","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EventDto"}}}}},"summary":"Get an event","tags":["events"]},"patch":{"operationId":"updateEvent","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateEventDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EventDto"}}}}},"summary":"Update an event","tags":["events"]},"delete":{"operationId":"deleteEvent","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"summary":"Delete an event","tags":["events"]}},"/events/{id}/signup":{"post":{"operationId":"signupToEvent","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignupToEventDto"}}}},"responses":{"201":{"description":""}},"summary":"Signup to an event","tags":["events"]}},"/events/{id}/members":{"get":{"operationId":"getEventMembers","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/EventMemberWithUser"}}}}}},"summary":"Get event members","tags":["events"]}},"/events/{id}/members/{memberId}":{"get":{"operationId":"getEventMember","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}},{"name":"memberId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EventMemberDto"}}}}},"summary":"Get an event member","tags":["events"]}},"/photos":{"post":{"operationId":"createPhotos","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CreatePhotoDto"}}}}},"responses":{"200":{"description":""},"201":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}},"summary":"Create a new photos","tags":["photos"]},"get":{"operationId":"getPhotos","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PhotoDto"}}}}}},"summary":"Get photos","tags":["photos"]},"delete":{"operationId":"deleteAllPhotos","parameters":[],"responses":{"200":{"description":""}},"summary":"Delete all photos","tags":["photos"]}},"/photos/unused":{"get":{"operationId":"getUnusedPhotos","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PhotoDto"}}}}}},"summary":"Get unused photos","tags":["photos"]}},"/photos/unused/next":{"get":{"operationId":"getNextUnusedPhoto","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PhotoDto"}}}}},"summary":"Get next unused photo","tags":["photos"]}},"/photos/unused/count":{"get":{"operationId":"countUnusedPhotos","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"number"}}}}},"summary":"Count unused photos","tags":["photos"]}},"/photos/{id}":{"get":{"operationId":"getPhoto","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PhotoDto"}}}}},"summary":"Get photo","tags":["photos"]},"patch":{"operationId":"updatePhoto","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdatePhotoDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PhotoDto"}}}}},"summary":"Update photo","tags":["photos"]}},"/settings":{"get":{"operationId":"getSettings","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SettingsDto"}}}}},"summary":"Get settings","tags":["Settings"]},"patch":{"operationId":"updateSettings","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateSettingsDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SettingsDto"}}}}},"summary":"Update settings","tags":["Settings"]}},"/categories":{"post":{"operationId":"createCategory","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCategoryDto"}}}},"responses":{"201":{"description":"Category data","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CategoryDto"}}}}},"summary":"Create a new category","tags":["categories"]},"get":{"operationId":"getAllCategories","parameters":[],"responses":{"200":{"description":"List of categories","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CategoryDto"}}}}}},"summary":"Get all categories","tags":["categories"]}},"/categories/schedule":{"get":{"operationId":"getSchedule","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CategoryScheduleDto"}}}}}},"summary":"Get schedule","tags":["categories"]},"patch":{"operationId":"updateSchedule","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateScheduleDto"}}}},"responses":{"200":{"description":"Updated schedule","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CategoryScheduleDto"}}}}},"summary":"Update schedule","tags":["categories"]}},"/categories/{id}/schedule":{"get":{"operationId":"getCategorySchedule","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CategoryScheduleDto"}}}}}},"summary":"Get category schedule","tags":["categories"]}},"/categories/{id}":{"get":{"operationId":"getCategory","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"Category data","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CategoryDto"}}}}},"summary":"Get category","tags":["categories"]},"patch":{"operationId":"updateCategory","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCategoryDto"}}}},"responses":{"200":{"description":"Category data","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CategoryDto"}}}}},"summary":"Update category","tags":["categories"]},"delete":{"operationId":"deleteCategory","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"summary":"Delete category","tags":["categories"]}}},"info":{"title":"Druzhinarus Chatbot REST API","description":"API specification","version":"0.1.0","contact":{}},"tags":[],"servers":[{"url":"http://localhost:7000/api","description":"Dev Server"}],"components":{"schemas":{"CreateUserDto":{"type":"object","properties":{"fullname":{"type":"string"},"email":{"type":"string","format":"email"},"phone":{"type":"string"},"birthday":{"format":"date-time","type":"string"},"vk":{"type":"string"}},"required":["fullname"]},"CreateTelegramAccountDto":{"type":"object","properties":{"channelId":{"type":"number"},"username":{"type":"string"},"user":{"$ref":"#/components/schemas/CreateUserDto"}},"required":["channelId","user"]},"UserDto":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"birthday":{"type":"string","nullable":true,"format":"date-time"},"role":{"type":"string","enum":["OWNER","ADMIN","MEMBER","USER"]},"fullname":{"type":"string"},"phone":{"type":"string","nullable":true},"email":{"type":"string","nullable":true},"vk":{"type":"string","nullable":true}},"required":["id","birthday","role","fullname","phone","email","vk"]},"AccountWithUserDto":{"type":"object","properties":{"id":{"type":"number"},"type":{"type":"string","enum":["TELEGRAM","VK"]},"channelId":{"type":"number"},"username":{"type":"string"},"user":{"$ref":"#/components/schemas/UserDto"}},"required":["id","type","channelId","user"]},"UpdateUserDto":{"type":"object","properties":{"fullname":{"type":"string"},"email":{"type":"string","format":"email"},"phone":{"type":"string"},"birthday":{"format":"date-time","type":"string"},"vk":{"type":"string"}}},"CreateEventDto":{"type":"object","properties":{"type":{"type":"string","enum":["COMMON","TRIP"]},"title":{"type":"string"},"startDate":{"format":"date-time","type":"string"},"endDate":{"format":"date-time","type":"string"},"link":{"type":"string"},"isPublic":{"type":"boolean"},"price":{"type":"number"},"discountedPrice":{"type":"number"},"discountEndDate":{"format":"date-time","type":"string"}},"required":["type","title","startDate"]},"EventDto":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"type":{"type":"string","enum":["COMMON","TRIP"]},"startDate":{"type":"string","format":"date-time"},"endDate":{"type":"string","nullable":true,"format":"date-time"},"price":{"type":"integer","nullable":true,"format":"int32"},"discountedPrice":{"type":"integer","nullable":true,"format":"int32"},"discountEndDate":{"type":"string","nullable":true,"format":"date-time"},"title":{"type":"string"},"link":{"type":"string","nullable":true},"isPublic":{"type":"boolean"}},"required":["id","type","startDate","endDate","price","discountedPrice","discountEndDate","title","link","isPublic"]},"EventMemberWithUser":{"type":"object","properties":{"eventId":{"type":"integer","format":"int32"},"userId":{"type":"integer","format":"int32"},"user":{"$ref":"#/components/schemas/UserDto"}},"required":["eventId","userId","user"]},"EventWithMembers":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"type":{"type":"string","enum":["COMMON","TRIP"]},"startDate":{"type":"string","format":"date-time"},"endDate":{"type":"string","nullable":true,"format":"date-time"},"price":{"type":"integer","nullable":true,"format":"int32"},"discountedPrice":{"type":"integer","nullable":true,"format":"int32"},"discountEndDate":{"type":"string","nullable":true,"format":"date-time"},"members":{"type":"array","items":{"$ref":"#/components/schemas/EventMemberWithUser"}},"title":{"type":"string"},"link":{"type":"string","nullable":true},"isPublic":{"type":"boolean"}},"required":["id","type","startDate","endDate","price","discountedPrice","discountEndDate","members","title","link","isPublic"]},"UpdateEventDto":{"type":"object","properties":{"title":{"type":"string"},"type":{"type":"object"},"startDate":{"format":"date-time","type":"string"},"endDate":{"format":"date-time","type":"string"},"link":{"type":"string"},"isPublic":{"type":"boolean"}}},"SignupToEventDto":{"type":"object","properties":{"userId":{"type":"number"}},"required":["userId"]},"EventMemberDto":{"type":"object","properties":{"eventId":{"type":"integer","format":"int32"},"userId":{"type":"integer","format":"int32"}},"required":["eventId","userId"]},"CreatePhotoDto":{"type":"object","properties":{"fileId":{"type":"string"},"groupId":{"type":"string"}},"required":["fileId"]},"PhotoDto":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"fileId":{"type":"string"},"groupId":{"type":"string","nullable":true},"isUsed":{"type":"boolean"}},"required":["id","fileId","groupId","isUsed"]},"UpdatePhotoDto":{"type":"object","properties":{"isUsed":{"type":"boolean"}},"required":["isUsed"]},"SettingsDto":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"photoSchedulerSpec":{"type":"string"}},"required":["id","photoSchedulerSpec"]},"UpdateSettingsDto":{"type":"object","properties":{}},"CreateCategoryDto":{"type":"object","properties":{"title":{"type":"string"},"description":{"type":"string"}},"required":["title","description"]},"CategoryDto":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"title":{"type":"string"},"description":{"type":"string"}},"required":["id","createdAt","updatedAt","title","description"]},"CategoryScheduleDto":{"type":"object","properties":{"weekday":{"type":"integer","format":"int32"},"hour":{"type":"integer","format":"int32"},"categoryId":{"type":"integer","nullable":true,"format":"int32"}},"required":["weekday","hour","categoryId"]},"UpdateScheduleDto":{"type":"object","properties":{"weekday":{"type":"number"},"hour":{"type":"number"},"categoryId":{"type":"number","nullable":true}},"required":["weekday","hour","categoryId"]},"UpdateCategoryDto":{"type":"object","properties":{"title":{"type":"string"},"description":{"type":"string"}}}}}}